Optimize Dockerfile for readability and maintainability
Streamlined the Dockerfile to improve readability by standardizing the case of 'AS' in multistage build keywords. Removed the installation of superfluous packages to prevent unnecessary layers and reduce the image size. Additionally, version logging into /VERSION is commented out as it's potentially being moved to a more suitable stage of the build process or an external script for clarity.