Optimize poetry installation in Dockerfile
Revised the installation steps for poetry in the Dockerfile. Removed usage of pipx and switched to using pip for installing and upgrading poetry directly. This simplifies the install process and reduces the layering within the Docker image. The change also prevents potential issues related to pipx path configurations, streamlining the container build process.