Updated Poetry Dependencies for Huggingface Support
Enhanced our Docker container by adding the 'embeddings-huggingface' extra in the Poetry install step, ensuring compatibility with HuggingFace models. This update allows the use of a wider range of pre-trained models and is expected to improve the versatility of our NLP features. Removed commented cleanup steps to simplify the Dockerfile.

Resolves #1234 (if an issue about adding Huggingface support was present)